platform/pyproject.toml
[project]
name = "contextagora"
version = "0.1.10"
description = "ContextAgora — context management for AI agents"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "python-multipart",
    "uvicorn",
    "httpx",
    "pyyaml>=6.0.3",
    "pydantic-settings",
]

[project.scripts]
start = "src.server:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-playwright>=0.7.2",
]
